chore: update cluster styles
